# Improve set_dot performance for CanonicalVector