Benchmark before
```
  835.672 μs (12 allocations: 7.65 MiB)
  849.430 μs (12 allocations: 7.65 MiB)
```
After
```
  3.437 ns (0 allocations: 0 bytes)
  6.182 ns (0 allocations: 0 bytes)
```